Add files via upload
This notebook creates the embeddings for all the BERTs models and then tries to evaluate them. So far, optimization is required on evaluation tasks or additional computational power. I think that a clever way is to separate the evaluation in another file and test on the generated and saved embeddings